requirements/base.txt
pytz==2018.9  # https://github.com/stub42/pytz
python-slugify==3.0.0  # https://github.com/un33k/python-slugify
Pillow==5.4.1  # https://github.com/python-pillow/Pillow
argon2-cffi==19.1.0  # https://github.com/hynek/argon2_cffi
redis==3.2.1  # https://github.com/antirez/redis
celery==4.2.1  # pyup: < 5.0  # https://github.com/celery/celery
kombu==4.4.0  # https://github.com/celery/kombu

# Django
# ------------------------------------------------------------------------------
django==2.0.13  # pyup: < 2.1  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==3.1.2  # https://github.com/jazzband/django-model-utils
django-allauth==0.39.1  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.7.2  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.10.0  # https://github.com/niwinz/django-redis

# Django REST Framework
djangorestframework==3.9.2  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client


distro==1.4.0
numpy==1.16.2
pandas==0.24.1
python-dateutil==2.8.0
pytz==2018.9
six==1.12.0
tabula-py==1.3.1
urllib3==1.24.1
PyPDF2==1.26.0
xlrd==1.2.0

django-jet==1.0.8
django-prettyjson==0.4.1
